+++ tests/selenium/reconnect/test_edge.py
@@ -6,24 +6,28 @@ from OnlySnarf.util.config import config
 from OnlySnarf.lib.driver import Driver
 from OnlySnarf.util.settings import Settings
 
-class TestSeleniumBrowsers(unittest.TestCase):
+class TestSeleniumReconnectEdge(unittest.TestCase):
 
     def setUp(self):
         config["debug_selenium"] = True
-        config["keep"] = False
+        config["keep"] = True
         # config["show"] = True
         Settings.set_debug("tests")
         self.driver = Driver()
 
     def tearDown(self):
         config["debug_selenium"] = False
+        config["keep"] = False
         config["show"] = False
         self.driver.exit()
 
-    def test_auto(self):
+    def test_reconnect_edge(self):
+        config["browser"] = "edge"
+        self.driver.init()
+        self.driver.exit()
         config["browser"] = "auto"
         self.driver.init()
-        assert self.driver.browser, "unable to launch via auto"
+        assert self.driver.browser, "unable to launch via reconnect edge"
 
 ############################################################################################